refactor(modal): rename scss files to use .module.scss
