+++ draft-packages/modal/KaizenDraft/Modal/Presets/ContextModal.module.scss
@@ -3,7 +3,7 @@
 @import "~@kaizen/design-tokens/sass/spacing";
 
 .modal {
-  composes: genericModal from "../Primitives/GenericModal.scss";
+  composes: genericModal from "../Primitives/GenericModal.module.scss";
   min-width: 300px;
   max-width: 840px;
