+++ draft-packages/modal/KaizenDraft/Modal/Primitives/ModalAccessibleLabel.tsx
@@ -1,7 +1,7 @@
+import React from "react"
 import classnames from "classnames"
-import * as React from "react"
 import { ModalAccessibleContext } from "./ModalAccessibleContext"
-import styles from "./ModalAccessibleLabel.scss"
+import styles from "./ModalAccessibleLabel.module.scss"
 
 export interface ModalAccessibleLabelProps {
   readonly children: React.ReactNode
