--- draft-packages/modal/KaizenDraft/Modal/Primitives/ModalBody.scss
@@ -1,24 +1,0 @@
-@import "~@kaizen/design-tokens/sass/color";
-@import "~@kaizen/deprecated-component-library-helpers/styles/type";
-
-.modalBody {
-  overflow-x: hidden;
-}
-
-.modalBody.scrollable {
-  overflow-y: scroll;
-}
-
-.modalBody.fillSpace {
-  flex-grow: 1;
-  min-height: 0;
-  overflow-y: auto;
-}
-
-.white {
-  background-color: $color-white;
-}
-
-.stone {
-  background-color: $color-gray-100;
-}
